Protect card edit and delete routes...
Only allow the user who created the card to edit or delete it.
In the future I as the superuser will also be able to edit or
delete a card. This is done using middleware that runs after
passport authentication that checks the card's new "contributor"
field user ID against the logged in user who sent the request.

Add an early version of user profile page to the client.