Redesign playground page layout and add some new content.
If not logged in (or if user decides to do so) the bonus can be
based upon hypothetical category spend instead of the actual
transactions.

Cards that are selected go to a separate window so it's clear what
cards are in the bonus calculation.

New bonus report includes:
- What currencies can be earned and in what quantity.
- Bonus value earned per transaction

Fix some bugs related to auth.