Add transaction table to bonus report
This table will show the bonus earned for every transaction and
which of the selected cards is ideally used for the transaction.

Also fix a bug where if card contributor is empty there's a crash.
It is able to use the default contributor by merging default
state with returned state from server.